Deprecate Valuer interface, use Rule interface instead (#18)
Deprecate the `Valuer` interface for literals and field values and use
the `Rule` interface instead. The signature is very similar, but now
that `Result` contains an `Error` field we can use it in place of
`Valuer`'s `ok bool` return value.

Also remove the `*notZero` node as we can now simply use the value
directly instead.